+++ Firestore/core/src/api/firestore.h
@@ -21,9 +21,9 @@
 #include <mutex>
 #include <string>
 
-#include "Firestore/core/interfaceForSwift/api/PipelineSource.h"
 #include "Firestore/core/src/api/api_fwd.h"
 #include "Firestore/core/src/api/load_bundle_task.h"
+#include "Firestore/core/src/api/pipeline_source.h"
 #include "Firestore/core/src/api/settings.h"
 #include "Firestore/core/src/core/core_fwd.h"
 #include "Firestore/core/src/credentials/credentials_fwd.h"
