+++ Firestore/core/src/api/pipeline.h
@@ -12,40 +12,49 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef FIRESTORE_CORE_INTERFACEFORSWIFT_API_PIPELINERESULT_H_
-#define FIRESTORE_CORE_INTERFACEFORSWIFT_API_PIPELINERESULT_H_
+#ifndef FIRESTORE_CORE_SRC_API_PIPELINE_H_
+#define FIRESTORE_CORE_SRC_API_PIPELINE_H_
 
+#include <functional>
 #include <memory>
+#include <vector>
+#include "Firestore/core/src/api/pipeline_result.h"
+#include "Firestore/core/src/api/stage.h"
 
 namespace firebase {
-
-class Timestamp;
-
 namespace firestore {
+
+namespace core {
+template <typename T>
+class EventListener;
+}  // namespace core
 
 namespace api {
 
 class Firestore;
-class DocumentReference;
+class PipelineResult;
 
-class PipelineResult {
+using PipelineSnapshotListener =
+    std::shared_ptr<core::EventListener<PipelineResult>>;
+
+class Pipeline {
  public:
-  PipelineResult(std::shared_ptr<Firestore> firestore,
-                 std::shared_ptr<Timestamp> execution_time,
-                 std::shared_ptr<Timestamp> update_time,
-                 std::shared_ptr<Timestamp> create_time);
+  Pipeline(std::shared_ptr<Firestore> firestore, Stage stage);
 
-  static PipelineResult GetTestResult(std::shared_ptr<Firestore> firestore);
+  void GetPipelineResult(PipelineSnapshotListener callback) const;
+
+  std::shared_ptr<Firestore> GetFirestore() const {
+    return firestore_;
+  }
 
  private:
   std::shared_ptr<Firestore> firestore_;
-  std::shared_ptr<Timestamp> execution_time_;
-  std::shared_ptr<Timestamp> update_time_;
-  std::shared_ptr<Timestamp> create_time_;
+  Stage stage_;
 };
 
 }  // namespace api
 
 }  // namespace firestore
 }  // namespace firebase
-#endif  // FIRESTORE_CORE_INTERFACEFORSWIFT_API_PIPELINERESULT_H_
+
+#endif  // FIRESTORE_CORE_SRC_API_PIPELINE_H_
