+++ Firestore/core/src/api/stage.cc
@@ -12,26 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef FIRESTORE_CORE_INTERFACEFORSWIFT_API_FIRESTOREPIPELINE_H_
-#define FIRESTORE_CORE_INTERFACEFORSWIFT_API_FIRESTOREPIPELINE_H_
-
-#include <memory>
-
-#include "PipelineSource.h"
+#include "Firestore/core/src/api/stage.h"
 
 namespace firebase {
 namespace firestore {
 
 namespace api {
-class Firestore;
 
-class FirestorePipeline {
- public:
-  static PipelineSource pipeline(std::shared_ptr<Firestore> firestore);
-};
+Stage::Stage() {
+}
 
 }  // namespace api
+
 }  // namespace firestore
 }  // namespace firebase
-
-#endif  // FIRESTORE_CORE_INTERFACEFORSWIFT_API_FIRESTOREPIPELINE_H_
